requirements.txt
asgiref==3.5.2
Django==4.1.4
django-filter==22.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
